chore: Separate review assignments on Review Page
